Fix playlist-utils (#980)
* dont standardize qualities for everything

* bump all ext depending directly or indirectly on playlist-utils
